refactored previous implementation
