+++ waterloop-site/src/testProfileData.js
@@ -1,19 +1,15 @@
-import LeadGooseOne from './images/LeadGoose.png';
-import LeadGithub from './images/BigGithub.png';
-import LeadLinkedIn from './images/BigLinkedIn.png';
-
 export default [
   {
     name: "A Potentially Very Long Name 1",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
@@ -21,14 +17,14 @@ export default [
   {
     name: "A Potentially Very Long Name 2",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
@@ -36,14 +32,14 @@ export default [
   {
     name: "A Potentially Very Long Name 3",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
@@ -51,14 +47,14 @@ export default [
   {
     name: "A Potentially Very Long Name 4",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
@@ -66,14 +62,14 @@ export default [
   {
     name: "A Potentially Very Long Name 5",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
@@ -81,14 +77,14 @@ export default [
   {
     name: "A Potentially Very Long Name 6",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
@@ -96,14 +92,14 @@ export default [
   {
     name: "A Potentially Very Long Name 7",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
@@ -111,14 +107,14 @@ export default [
   {
     name: "A Potentially Very Long Name 8",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
@@ -126,14 +122,14 @@ export default [
   {
     name: "A Potentially Very Long Name 9",
     position: "Papa Goose is the Alpha of V1 Green",
-    portrait: LeadGooseOne,
+    portrait: "/LeadGoose.png",
     contacts: [
       {
-        icon: LeadGithub,
+        icon: "/BigGithub.png",
         url: 'http://www.github.com'
       },
       {
-        icon: LeadLinkedIn,
+        icon: "/BigLinkedIn.png",
         url: 'http://www.linkedin.com'
       }
     ]
